refactor: move transaction/mod.rs tests into transaction/tests.rs
